third_party/blink/web_tests/platform/win11-arm64/http/tests/inspector-protocol/network/interception-take-stream-expected.txt

Tests that resource bodies can be taken as streams.

Running test: testBasicUsage
data: <br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\htt (100) eof: false
eof: false

Running test: testLargeRead
data: (100001) eof: false
Data mistmatch near offset 0, got <br />
<b>Warning</b>:  Undefined array key "type" in , expected: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789
eof: false

Running test: testSmallNonOverlappingReads
data: (13808)
Data mistmatch near offset 0, got <br />
<b>Warning</b>:  Undefined array key "type" in , expected: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789

Running test: testSmallOverlappingReads
data: (103808)
Data mistmatch near offset 0, got <br />
<b>Warning</b>:  Undefined array key "type" in , expected: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789

Running test: testRead0
reading 0: Invalid max read size
data: <br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\htt (100) eof: false

Running test: testTakeTwice
Trying to take stream twice: Can only get response body on requests captured after headers received.

Running test: testReadAfterCancel
data: <br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\htt (100) eof: false
eof: false

Running test: testContinueAfterBodyTaken
Attempting to continue as is after taking request: Unable to continue request as is after body is taken

Running test: testReadAfterClose
Attempting read after close: Invalid stream handle
